Hardware abstraction.
